+++ decidim-core/lib/decidim/content_parsers/user_group_parser.rb
@@ -2,51 +2,51 @@
 
 module Decidim
   module ContentParsers
-    # A parser that searches user mentions in content.
+    # A parser that searches user groups mentions in content.
     #
     # A word starting with `@` will be considered as a possible mention if
     # they only contains letters, numbers or underscores.
     #
     # @see BaseParser Examples of how to use a content parser
-    class UserParser < BaseParser
+    class UserGroupParser < BaseParser
       # Class used as a container for metadata
       #
-      # @!attribute users
-      #   @return [Array] an array of Decidim::User mentioned in content
-      Metadata = Struct.new(:users)
+      # @!attribute groups
+      #   @return [Array] an array of Decidim::UserGroup mentioned in content
+      Metadata = Struct.new(:groups)
 
       # Matches a nickname if contains letters, numbers or underscores.
       MENTION_REGEX = /\B@(\w*)\b/.freeze
 
       # Replaces found mentions matching a nickname of an existing
-      # user in the current organization with a global id. Other
-      # mentions found that doesn't match an existing user are
+      # group in the current organization with a global id. Other
+      # mentions found that doesn't match an existing group are
       # returned as is.
       #
       # @return [String] the content with the valid mentions replaced by a global id
       def rewrite
         content.gsub(MENTION_REGEX) do |match|
-          users[match[1..-1]]&.to_global_id&.to_s || match
+          groups[match[1..-1]]&.to_global_id&.to_s || match
         end
       end
 
       # (see BaseParser#metadata)
       def metadata
-        Metadata.new(existing_users)
+        Metadata.new(existing_groups)
       end
 
       private
 
-      def users
-        @users ||= Hash[
-          existing_users.map do |user|
-            [user.nickname, user]
+      def groups
+        @groups ||= Hash[
+          existing_groups.map do |group|
+            [group.nickname, group]
           end
         ]
       end
 
-      def existing_users
-        @existing_users ||= Decidim::User.where(organization: current_organization, nickname: content_nicknames)
+      def existing_groups
+        @existing_groups ||= Decidim::UserGroup.where(organization: current_organization, nickname: content_nicknames)
       end
 
       def content_nicknames